devmem: add --no-sync option.
Although O_SYNC is what you want 99% of the time, there is also a use
for not supplying the flag. See drivers/char/mem.c's uncached_access().

The big question here is probably the bikeshed issue of whether to name
the option after what it does (disable the O_SYNC flag) or _why_ you're
doing that. I opted for the former on the basis of "given the -f
argument, you're not necessarily talking to the /dev/mem driver, and
other drivers might use this for other things", even though I'm not
aware of any such driver personally. (Also, it's the shorter name!)