crond: Fix timing
I found a few more issues I wanted to fix, all related to environment
variables.
- When setting a variable in the crontab file, the value would have a
newline at the end of it. That's because getline() returns the line
with a newline at the end, and the code didn't strip that newline out.
- The HOME variable was being sourced from what HOME was set to when
crond was run, rather than what the user's home is in the passwd file.
I believe the intention was to set HOME to what it is in the passwd
file, but allow the crontab to override it.
- I decided to also set LOGNAME to the username of the crontab's
owner, since the posix manpage for crontab mentions that as one of the
variables that should be set.