crontab: Make sure crond detects crontab modifications
Currently crond uses the modification time of the crontabs directory
to detect changes to the crontab files to decide whether to reload
them. This works fine when crontab creates or removes crontab files,
but when an existing crontab file is modified or replaced, it doesn't
change the modification time of the directory, so crond doesn't know
it should reload the crontab files. This patch fixes this by making
crontab remove the old crontab file (if it exists) before recreating
it, so the directory's modification time always changes when a crontab
is modified.