pgrep: Implement -a
pgrep currently implements -l and -f, but not -a. It also inadvertently
treats -f as a combination of -a and -f, where we both filter via the
full command line and display the full command line. This violates
--help, where -f says "Check full command line for PATTERN".

Implement -a, which takes care of displaying the full command line, and
return -f to simply checking (filtering) via command line.