Add TOYFLAG_MOREHELP(CFG_BLAH) to allow annotated help text to drop out when a second config symbol isn't defined. Use this for various LSM -Z flags, PASSWD_SAD, sort -g, and wget's https support. This replaces the old help text merging scripts/config2help.c used to do.
The annotation is a leading !, which removes the next char from usage:
lines and the whole line from the rest of the help text. The ! is always
removed, and the data it marks is only shown if the argument to
TOYFLAG_MOREHELP() is true at compile time.

In theory this plumbing should drop out when not used, like lib/args.c.